Cleaned up solvers a bit more
