+++ Assets/MixedRealityToolkit/_Core/Utilities/Physics/Solvers/SolverHandler.cs
@@ -6,30 +6,54 @@ using UnityEngine;
 
 namespace Microsoft.MixedReality.Toolkit.Internal.Utilities.Physics.Solvers
 {
+    /// <summary>
+    /// This class handles the solver components that are attached to this <see cref="GameObject"/>
+    /// </summary>
     public class SolverHandler : MonoBehaviour
     {
         [SerializeField]
+        [Tooltip("The target transform that the solvers will act upon.")]
         private Transform transformTarget;
 
+        /// <summary>
+        /// The target transform that the solvers will act upon.
+        /// </summary>
         public Transform TransformTarget
         {
             get { return transformTarget; }
             set { transformTarget = value; }
         }
 
+        /// <summary>
+        /// The position the solver is trying to move to.
+        /// </summary>
         public Vector3 GoalPosition { get; private set; }
 
+        /// <summary>
+        /// The rotation the solver is trying to rotate to.
+        /// </summary>
         public Quaternion GoalRotation { get; private set; }
 
+        /// <summary>
+        /// The scale the solver is trying to scale to.
+        /// </summary>
         public Vector3 GoalScale { get; private set; }
 
+        /// <summary>
+        /// Alternate scale.
+        /// </summary>
         public Vector3 AltScale { get; private set; }
 
+        /// <summary>
+        /// The timestamp the solvers will use to calculate with.
+        /// </summary>
         public float DeltaTime { get; private set; }
 
         private float lastUpdateTime;
 
         private readonly List<Solver> solvers = new List<Solver>();
+
+        #region Monobehaviour Implementation
 
         private void OnValidate()
         {
@@ -67,5 +91,7 @@ namespace Microsoft.MixedReality.Toolkit.Internal.Utilities.Physics.Solvers
                 }
             }
         }
+
+        #endregion Monobehaviour Implementation
     }
 }
